refactor ui components
