+++ src/app/components/Footer.tsx
@@ -2,7 +2,7 @@
 import { Typography } from "@material-tailwind/react";
 import Image from "next/image";
 import { useContext } from "react";
-import { ThemeContext } from "./ThemeProvider";
+import { ThemeContext } from "./components/ThemeProvider";
 import SITEMAP from "./Sitemap";
 
 const currentYear = new Date().getFullYear();
