+++ src/app/components/Transition.tsx
@@ -1,0 +1,5 @@
+import { useTransition } from "react";
+
+export default function Transition(){
+    
+}
